Add onboarding flow, metric UI, and persistence
Introduce a full onboarding flow and supporting UI/components and persistence:

- Add onboarding screens: WelcomeScreen, IdentityScreen, MetricsScreen and integrate them into AegisNavigation (start destination depends on onboarding state). Hide bottom bar on onboarding routes and clear onboarding backstack on completion.
- Add MetricInput and AegisTextField components for metrics entry and a Welcome UI.
- Persist onboarding state: add onboarding_completed boolean to SettingsStore with flow and save function; expose onboardingCompleted in UserRepository and ProfileViewModel; add completeOnboarding() in ViewModel.
- Photo handling improvements: use PickVisualMedia in ProfileScreen/ProfileContent and IdentityScreen, persist URI permissions, and ensure avatar/base photo is synchronized and persisted via ProfileViewModel changes.
- Type and default fixes: change height handling to Double across viewmodel/repo/settings, and update default height from 1.70 to 0.0; adapt SettingsComponents to send Double.
- Add several theme colors (MatteBlack, AegisSteel, AegisCream) and small cleanup (remove unused imports, minor UI tweaks).

These changes wire up onboarding end-to-end (UI, navigation, DataStore persistence) and add improved metric input components and photo picker robustness.

+++ app/src/main/java/com/yago/aegis/ui/components/SettingsComponents.kt
@@ -114,7 +114,7 @@ fun SettingsMenu(viewModel: ProfileViewModel) {
             onValueChange = { newValue ->
                 if (newValue.all { it.isDigit() }) {
                     tempHeight = newValue
-                    newValue.toIntOrNull()?.let { viewModel.updateHeight(it) }
+                    newValue.toIntOrNull()?.let { viewModel.updateHeight(it.toDouble()) }
                 }
             },
             label = { Text(stringResource(R.string.label_height), color = Color.Gray) },
